+++ frontend/src/tables/UniversalTable.tsx
@@ -6,6 +6,8 @@ import DeleteDialog from "./elements/DeleteDialog";
 import EditDialog from "./elements/EditDialog";
 import UpdateRowButton from "./elements/UpdateRowButton";
 import TableHeader from "./elements/TableHeader";
+import ReadRowDialog from "./elements/ReadRowDialog";
+import ReadRowButton from "./elements/ReadRowButton";
 
 interface Props {
   useStore: ReturnType<typeof createTableStore>;
@@ -37,23 +39,25 @@ export default function UniversalTable({ useStore }: Props) {
               key={row[data.table_options.pk]?.toString()}
               style={{ textAlign: "center", verticalAlign: "middle" }}
             >
-              {/* edit button cell */}
-              {data.table_options.update__ask_columns.length ? (
-                <td
-                  key={row[data.table_options.pk]?.toString() + "__edit"}
-                  style={{ width: "1px" }}
-                >
+              {/* view/edit button cell */}
+              <td
+                key={row[data.table_options.pk]?.toString() + "__edit"}
+                style={{ width: "1px" }}
+              >
+                {data.table_options.update__ask_columns.length ? (
                   <UpdateRowButton
                     row={row}
                     useStore={useStore}
                   ></UpdateRowButton>
-                </td>
-              ) : null}
+                ) : (
+                  <ReadRowButton row={row} useStore={useStore}></ReadRowButton>
+                )}
+              </td>
               {/* visible columns */}
               {tableStore.data?.table_options?.read__visible_columns.map(
                 (col) => (
                   <ViewCellElement
-                    data={data}
+                    useStore={useStore}
                     row={row}
                     col={col}
                     key={
@@ -81,6 +85,7 @@ export default function UniversalTable({ useStore }: Props) {
           ))}
         </tbody>
       </Table>
+      <ReadRowDialog useStore={useStore}></ReadRowDialog>
       <EditDialog useStore={useStore}></EditDialog>
       <DeleteDialog useStore={useStore}></DeleteDialog>
     </>